connection config, and changes in data extraction loop

+++ SP/2.USP_GM_DataExtraction.sql
@@ -199,9 +199,21 @@ Set @FailPoint=3
 				END TRY
 
 				BEGIN CATCH
-			
-					TRUNCATE TABLE #TEMP_RawData
-					SET @Sucsess = 0
+
+					IF @ErrorMessage not like '%Object reference not set to an instance of an object%' and @ErrorMessage not like '%No connection could be made because the target machine actively refused it%'
+					BEGIN
+							EXEC AdvancedBIsystem.dbo.USP_GAL_InsertLogEvent @LogEventObjectName = 'USP_GM_DataExtraction', @EngineName = 'MFG_Solutions', 
+							@ModuleName = 'DataExtraction', @LogEventMessage = @ErrorMessage, @LogEventType = 'E' 
+							RAISERROR (N'USP_GM_DataExtraction::FailPoint- %d ERR-%s', 16,1, @FailPoint, @ErrorMessage)
+					END
+
+					ELSE BEGIN					
+						TRUNCATE TABLE #TEMP_RawData
+						SET @Sucsess = 0
+						SET @ErrorMessage = 'Fail Point: ' + CONVERT(VARCHAR(3), @FailPoint) +' '+ ERROR_MESSAGE()
+						SELECT @ErrorMessage
+					END
+					--SELECT GETUTCDATE(),'select * from('+@ImportQuery+')Q where '+@DistributionField+'%'+convert(varchar(max),@NumDistributionGroups)+'='+convert(varchar(max),@i)
 
 				END CATCH
 			
@@ -297,18 +309,18 @@ BEGIN CATCH
 
 SET @ErrorMessage = 'Fail Point: ' + CONVERT(VARCHAR(3), @FailPoint) +' '+ ERROR_MESSAGE()
 SELECT @ErrorMessage
-IF @ErrorMessage not like '%Object reference not set to an instance of an object%' and @ErrorMessage not like '%No connection could be made because the target machine actively refused it%' and @NumRecoveryRetry<7
-GOTO OTHER_ERROR
-	INSERT INTO dbo.IBI_FailingQueries(UTC_TimeStamp,Query)
-	SELECT GETUTCDATE(),'select * from('+@ImportQuery+')Q where '+@DistributionField+'%'+convert(varchar(max),@NumDistributionGroups)+'='+convert(varchar(max),@i)
+--IF @ErrorMessage not like '%Object reference not set to an instance of an object%' and @ErrorMessage not like '%No connection could be made because the target machine actively refused it%' and @NumRecoveryRetry<7
+--GOTO OTHER_ERROR
+	--INSERT INTO dbo.IBI_FailingQueries(UTC_TimeStamp,Query)
+	--SELECT GETUTCDATE(),'select * from('+@ImportQuery+')Q where '+@DistributionField+'%'+convert(varchar(max),@NumDistributionGroups)+'='+convert(varchar(max),@i)
 
-	SET @NumRecoveryRetry=ISNULL(@NumRecoveryRetry+1,1)
-	PRINT('Recovering from fail: '+convert(varchar(100),@NumRecoveryRetry))
-	EXEC dbo.USP_GM_DataExtraction @ModelGroupID=@ModelGroupID,@NumRecoveryRetry=@NumRecoveryRetry,@i=@i,@RecoveryQueryNum=@QueryNum,@LogMessage=@LogMessage
-	GOTO FINISH
-OTHER_ERROR:
+	--SET @NumRecoveryRetry=ISNULL(@NumRecoveryRetry+1,1)
+	--PRINT('Recovering from fail: '+convert(varchar(100),@NumRecoveryRetry))
+	--EXEC dbo.USP_GM_DataExtraction @ModelGroupID=@ModelGroupID,@NumRecoveryRetry=@NumRecoveryRetry,@i=@i,@RecoveryQueryNum=@QueryNum,@LogMessage=@LogMessage
+	--GOTO FINISH
+--OTHER_ERROR:
 	EXEC AdvancedBIsystem.dbo.USP_GAL_InsertLogEvent @LogEventObjectName = 'USP_GM_DataExtraction', @EngineName = 'MFG_Solutions', 
 							@ModuleName = 'DataExtraction', @LogEventMessage = @ErrorMessage, @LogEventType = 'E' 
 	RAISERROR (N'USP_GM_DataExtraction::FailPoint- %d ERR-%s', 16,1, @FailPoint, @ErrorMessage)
-FINISH:
+--FINISH:
 END CATCH
